% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text_to_vector.R
\name{add_text_to_vector}
\alias{add_text_to_vector}
\title{add text to a vector}
\usage{
add_text_to_vector(the_vector, the_text, override_NA = TRUE, sep = ", ")
}
\arguments{
\item{the_vector}{Specify a vector that should be edited.}

\item{the_text}{Character. Specify the text you want to add to your vector.}

\item{override_NA}{Logical. Choose TRUE if you want to override NA values.}

\item{sep}{Character. Specify the separator between the text.}
}
\description{
This function will add predefined text to a specified vector.
  The special aspect is that it will not repeat the text if it exists.
}
\note{
The idea behind the function is that you can specify a subset of a data frame and add values to it, without overriding existing values. It is meant to prepare a data frame before calling badval_NA.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_NA}}
}

