% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badindex_add.R
\name{badindex_add}
\alias{badindex_add}
\title{Add bad values}
\usage{
badindex_add(data, subset, badindex = "BADVAL", badstring,
  override_NA = TRUE, separator = ", ")
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{subset}{Integer. A vector specifying the rows that contain bad
values.}

\item{badindex}{Character. Name of a column of the data frame that serves as an
index for bad values.}

\item{badstring}{Character. Specify the text for indexing (i.e. the names of columns that contain bad values).}

\item{override_NA}{Logical. Choose TRUE if you want to override NA values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\value{
'badindex' column (vector).
}
\description{
This function will add predefined text to a specified vector.
  The special aspect is that it will not repeat the text if it exists.
}
\note{
The idea behind the function is that you can specify a subset of a data
  frame and add values to it, without overriding existing values. It is meant
  to prepare the data frame before calling badvalue_rm.
}
\seealso{
\code{\link{badvalue_rm}}, \code{\link{badindex_rm}}
}
\author{
Frederik Sachser
}
