% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badindex_rm.R
\name{badindex_rm}
\alias{badindex_rm}
\title{Remove bad values from the index}
\usage{
badindex_rm(data, subset, badstring, badindex = "BADVAL", separator = ", ")
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{subset}{Integer. A vector specifying the rows that contain bad
values.}

\item{badstring}{Character. Specify the text for indexing (i.e. the names of columns that contain bad values).}

\item{badindex}{Character. Name of a column of the data frame that serves as an
index for bad values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\description{
This function will remove predefined text from a specified vector.
}
\seealso{
\code{\link{badvalue_rm}}, \code{\link{badindex_add}}
}
\author{
Frederik Sachser
}
