% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badindex_rm_na.R
\name{badindex_rm_na}
\alias{badindex_rm_na}
\title{Remove badstring if is.na(badvalue)}
\usage{
badindex_rm_na(data, badstring = NULL, badindex = "BADVAL",
  separator = ",")
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{badstring}{Character. Specify the text for indexing (i.e. the names of columns that contain bad values).}

\item{badindex}{Character. Name of a column of the data frame that serves as an
index for bad values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\description{
This function will remove corresponding badstrings from badindex if badvalues are missing (NA).
}
\seealso{
\code{\link{badindex_rm_sep}}
}
\author{
Frederik Sachser
}
