% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badindex_rm_sep.R
\name{badindex_rm_sep}
\alias{badindex_rm_sep}
\title{Clean badindex - remove unnecessary separators}
\usage{
badindex_rm_sep(data, badindex = "BADVAL", separator = ", ")
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{badindex}{Character. Name of a column of the data frame that serves as an
index for bad values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\description{
This function will remove unnecessary separators from badindex.
}
\note{
The function won´t work with spaces in column names.
}
\seealso{
\code{\link{badstring_exist}}
}
\author{
Frederik Sachser
}
