% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badstring_exist.R
\name{badstring_exist}
\alias{badstring_exist}
\title{Check badindex - find misspelled badstrings.}
\usage{
<<<<<<< HEAD
badstring_exist(data, badindex, separator = ", ")
=======
badstring_exist(data, badindex = "BADVAL", separator = ", ")
>>>>>>> rewrite
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{badindex}{Character. Name of a column of the data frame that serves as an
index for bad values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\description{
This function will check for entries not matched in colnames (badstrings).
}
\note{
The function won´t work with spaces in column names.
}
\seealso{
\code{\link{badindex_rm_sep}}
}
\author{
Frederik Sachser
}
