% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_add.R
\name{badval_add}
\alias{badval_add}
\title{Add bad values}
\usage{
badval_add(badval_column, bad_row_index, bad_col_name, override_NA = TRUE,
  separator = ", ")
}
\arguments{
\item{badval_column}{A vector (or column in a data frame) that serves as an
index for bad values.}

\item{bad_row_index}{Integer. A vector specifying the rows that contain bad
values.}

\item{bad_col_name}{Character. Specify the text you want to add to your
vector (i.e. the names of columns that contain bad values).}

\item{override_NA}{Logical. Choose TRUE if you want to override NA values.}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\value{
badval_column
}
\description{
This function will add predefined text to a specified vector.
  The special aspect is that it will not repeat the text if it exists.
}
\note{
The idea behind the function is that you can specify a subset of a data
  frame and add values to it, without overriding existing values. It is meant
  to prepare the data frame before calling badval_clean_data
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_clean_data}}, \code{\link{badval_rm_index}}
}

