% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_backup.R
\name{badval_backup}
\alias{badval_backup}
\title{Backup bad values}
\usage{
badval_backup(data, badval_column, badval_pattern, badval_exactly)
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{badval_column}{A vector (or column in a data frame) that serves as an
index for bad values.}

\item{badval_pattern}{Character. Specify a pattern of bad values to clean
multiple (but not all) columns of your data frame.}

\item{badval_exactly}{Character. Specify an exact string to clean only one
single column of your data frame.}
}
\value{
data frame
}
\description{
This function will backup the specified bad values of a data
  frame.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_restore}}
}

