% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_NA.R
\name{badval_NA}
\alias{badval_NA}
\title{Remove bad values from data}
\usage{
badval_NA(data, bad_values)
}
\arguments{
\item{data}{A data frame.}

\item{bad_values}{Character. A Vector of the same length as nrow(data).
Generally it is meant to be a column of the data where you can specify the
names of columns that has a bad value in this row. Therefore you do not need to
delete the cells with bad values in a data frame but specify them in a
column. If there are multiple bad values per row the column names should be
separated by comma.}
}
\description{
This function will remove cells of a data frame, that have bad
  values.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{add_text_to_vector}}
}

