% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_rm_data.R
\name{badval_rm_data}
\alias{badval_rm_data}
\title{Remove bad values from data}
\usage{
badval_rm_data(data, badval_column, badval_pattern, badval_exactly)
}
\arguments{
\item{data}{A data frame containing a column with an index of bad values.}

\item{badval_column}{A vector (or column in a data frame) that serves as an
index for bad values.}

\item{badval_pattern}{Character. Specify a pattern of bad values to clean
multiple (but not all) columns of your data frame.}

\item{badval_exactly}{Character. Specify an exact string to clean only one
single column of your data frame.}
}
\value{
data frame
}
\description{
This function will remove cells of a data frame, that have bad
  values.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_add_index}}, \code{\link{badval_rm_index}}
}

