% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_rm_index.R
\name{badval_rm_index}
\alias{badval_rm_index}
\title{Remove bad values from the index}
\usage{
badval_rm_index(badval_column, bad_col_name, separator = ", ")
}
\arguments{
\item{badval_column}{A vector (or column in a data frame) that serves as an
index for bad values.}

\item{bad_col_name}{Character. Specify the text you want to add to your
vector (i.e. the names of columns that contain bad values).}

\item{separator}{Character. Specify the separator between the text. It is not
advisable to use '|'.}
}
\value{
badval_column
}
\description{
This function will remove predefined text from a specified vector.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_clean_data}}, \code{\link{badval_add}}
}

