% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badval_restore.R
\name{badval_restore}
\alias{badval_restore}
\title{Restore bad values}
\usage{
badval_restore(good_col, backup_data, cleaned_data)
}
\arguments{
\item{good_col}{Character. Name of a column that should be restored.}

\item{backup_data}{A data frame containing only bad values and NA.}

\item{cleaned_data}{A data frame containing data without bad values}
}
\value{
data frame
}
\description{
This function will backup the specified bad values of a data
  frame.
}
\note{
The Index will compare and use row.names of both datasets to restore
  the values. Therefore you should not use the function if you manually
  changed the row names of the cleaned data after creating the backup data.
}
\author{
Frederik Sachser
}
\seealso{
\code{\link{badval_backup}}
}

